# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
  database?: string;
}

export type AuthResponseUser = {
  id: string;
  name?: string;
  role: string;
};

export interface AuthResponse {
  user: AuthResponseUser;
}

export interface CollectionResource {
  value: unknown;
  revision?: number;
}

export type AuthenticatePinBody = {
  role: string;
  pin: string;
};

export type ChangePinBody = {
  currentPin: string;
  /**
   * @minLength 4
   * @maxLength 12
   */
  newPin: string;
};

export type PutCollectionResourceBody = {
  value: unknown;
};

export type GetReportSummaryParams = {
  from?: string;
  to?: string;
};

export type GetReportSummary200 = { [key: string]: unknown };

export type ListTables200Item = { [key: string]: unknown };

export type UpdateTableBody = { [key: string]: unknown };

export type UpdateTable200 = { [key: string]: unknown };

export type ListZones200Item = { [key: string]: unknown };

export type ListPurchaseOrders200Item = { [key: string]: unknown };

export type CreatePurchaseOrderBody = { [key: string]: unknown };

export type CreatePurchaseOrder201 = { [key: string]: unknown };

export type SettleTransactionBody = { [key: string]: unknown };

export type SettleTransaction201 = { [key: string]: unknown };

export type ListAuditEntriesParams = {
  /**
   * @minimum 1
   * @maximum 500
   */
  limit?: number;
};

export type ListAuditEntries200Item = { [key: string]: unknown };